+++ creator-node/package.json
@@ -42,7 +42,6 @@
     "@opentelemetry/semantic-conventions": "^1.5.0",
     "@opentelemetry/tracing": "^0.24.0",
     "@solana/web3.js": "1.31.0",
-    "JSONStream": "^1.3.5",
     "async-retry": "^1.3.3",
     "axios": "^0.19.2",
     "base64-url": "^2.3.3",
@@ -76,6 +75,7 @@
     "hashids": "2.2.10",
     "ioredis": "5.2.4",
     "jimp": "^0.6.1",
+    "JSONStream": "^1.3.5",
     "lodash": "4.17.21",
     "multer": "^1.4.0",
     "pg": "^8.0.3",
