+++ creator-node/test/ffmpeg.test.js
@@ -4,7 +4,9 @@ const assert = require('assert')
 const { libs } = require('@audius/sdk')
 const Utils = libs.Utils
 
-const { segmentFile, transcodeFileTo320, getFileInformation } = require('../src/ffmpeg')
+const { computeTranscodeDestinationPath, getAudioFileInformation } = require('./lib/helpers')
+
+const { segmentFile, transcodeFileTo320 } = require('../src/ffmpeg')
 
 describe('test segmentFile()', () => {
   /**
@@ -95,42 +97,6 @@ describe('test segmentFile()', () => {
     assert.deepStrictEqual(allSegmentsSet.size, 0)
   })
 })
-
-function computeTranscodeDestinationPath (fileDir, fileName) {
-  return path.resolve(fileDir, fileName.split('.')[0] + '-dl.mp3')
-}
-
-/**
- * Ignores milliseconds bc of inaccuracy in comparison due to how we transcode
- * e.g. Input file info shows:
- *    [mp3 @ 0x72367c0] Estimating duration from bitrate, this may be inaccurate
- *    Duration: 00:03:07.44, start: 0.000000, bitrate: 320 kb/s
- * and output file shows:
- *    Duration: 00:03:07.46, start: 0.023021, bitrate: 320 kb/s
- * Note these durations are the same after accounting for the start offset (not sure why thats there)
- */
-async function getAudioFileInformation (filePath) {
-  const info = await getFileInformation(filePath)
-  if (!info) throw new Error('Failed to get file information')
-
-  let duration = /Duration: (\d{2}:\d{2}:\d{2}\.\d{2})/g.exec(info.toString())
-  if (!duration) throw new Error('Failed to find file duration')
-  duration = duration[1].split('.')[0]
-
-  let metadata = {}
-  // Extract the metadata properties using a regular expression
-  const properties = /^\s{4}(\w+)\s+:(.+)/gm
-  let match
-  while ((match = properties.exec(info.toString())) !== null) {
-    metadata[match[1]] = match[2].trim()
-  }
-
-  return {
-    info,
-    duration,
-    metadata
-  }
-}
 
 describe('test transcodeFileTo320()', () => {
   const fileDir = __dirname
